+++ Source/main.cpp
@@ -155,21 +155,21 @@ int main(void)
 void CameraUpdate(Camera2D& camera, Player& player, Screen& screen, Map& map) {
     
     if(IsKeyDown(KEY_LEFT_CONTROL)) screen.unscalezoom += (GetMouseWheelMove() * 0.05f);
-    float screenscale = (screen.currWidth / screen.windowedWidth);
+    float screenscale = ((float)screen.currWidth / screen.windowedWidth);
     if (screen.unscalezoom > screen.maxzoom) screen.unscalezoom = screen.maxzoom;
     else if (screen.unscalezoom < screen.minzoom) screen.unscalezoom = screen.minzoom;
 
     camera.zoom = screen.unscalezoom * screenscale;     //Breaking down zoom into unscale zoom and screenscale ensures
                                                         //the same viewport is rendered independent of window size.
 
-    camera.offset = Vector2(screen.currWidth/2, screen.currHeight / 2);
+    camera.offset = Vector2(screen.currWidth / 2.f, screen.currHeight / 2.f);
     camera.target = player.position;
 
     Vector2 stw = GetScreenToWorld2D(Vector2(0,0), camera);
     if (stw.x < 0) camera.target.x = camera.offset.x / camera.zoom;
     if (stw.y < 0) camera.target.y = camera.offset.y / camera.zoom;
 
-    stw = GetScreenToWorld2D(Vector2(screen.currWidth, screen.currHeight), camera);
+    stw = GetScreenToWorld2D(Vector2(static_cast<float>(screen.currWidth), static_cast<float>(screen.currHeight)), camera);
     if (stw.x > map.tilesX * map.tileSize) camera.target.x = map.tilesX * map.tileSize - camera.offset.x / camera.zoom;
     if (stw.y > map.tilesY * map.tileSize) camera.target.y = map.tilesY * map.tileSize - camera.offset.y / camera.zoom;
 
@@ -187,7 +187,7 @@ void DrawGui(Screen& screen, Player& player) {
             DrawRectangle(screen.currWidth / 2 + (60 * i) - 290, screen.currHeight - 55, 40, 40, DARKBROWN);
         }
         if (player.inventory.slot[i] != nullptr) {
-            DrawTextureRec(*player.inventory.slot[i]->sprite, player.inventory.slot[i]->spriteSource, Vector2(screen.currWidth / 2 + (60 * i) - 290, screen.currHeight - 55), WHITE);
+            DrawTextureRec(*player.inventory.slot[i]->sprite, player.inventory.slot[i]->spriteSource, Vector2(static_cast<float>(screen.currWidth / 2 + (60 * i) - 290), static_cast<float>(screen.currHeight - 55)), WHITE);
             if (player.inventory.slot[i]->currStack > 1) {
                 int spacing = 0;                                //Might be better to show text as img, in order to indent it to the right.
                 if (player.inventory.slot[i]->currStack > 100) spacing = 2;     //Using log() would make the code simpler, but slower.
